Move network resources into ec2 group
