+++ apis/compute/v1alpha3/referencers.go
@@ -23,8 +23,8 @@ import (
 
 	"github.com/crossplane/crossplane-runtime/pkg/reference"
 
+	network "github.com/crossplane/provider-aws/apis/ec2/v1beta1"
 	"github.com/crossplane/provider-aws/apis/identity/v1beta1"
-	network "github.com/crossplane/provider-aws/apis/network/v1beta1"
 )
 
 // ResolveReferences of this EKSCluster
