+++ apis/ec2/network.go
@@ -14,9 +14,4 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package v1beta1 contains managed resources for AWS network services such as
-// VPC and Subnet.
-// +kubebuilder:object:generate=true
-// +groupName=network.aws.crossplane.io
-// +versionName=v1beta1
-package v1beta1
+package ec2
