+++ config/crd/ec2.aws.crossplane.io_securitygroups.yaml
@@ -6,7 +6,7 @@ metadata:
   annotations:
     controller-gen.kubebuilder.io/version: v0.2.4
   creationTimestamp: null
-  name: securitygroups.network.aws.crossplane.io
+  name: securitygroups.ec2.aws.crossplane.io
 spec:
   additionalPrinterColumns:
   - JSONPath: .status.conditions[?(@.type=='Ready')].status
@@ -24,7 +24,7 @@ spec:
   - JSONPath: .metadata.creationTimestamp
     name: AGE
     type: date
-  group: network.aws.crossplane.io
+  group: ec2.aws.crossplane.io
   names:
     categories:
     - crossplane
