+++ config/crd/ec2.aws.crossplane.io_subnets.yaml
@@ -6,7 +6,7 @@ metadata:
   annotations:
     controller-gen.kubebuilder.io/version: v0.2.4
   creationTimestamp: null
-  name: internetgateways.network.aws.crossplane.io
+  name: subnets.ec2.aws.crossplane.io
 spec:
   additionalPrinterColumns:
   - JSONPath: .status.conditions[?(@.type=='Ready')].status
@@ -21,26 +21,28 @@ spec:
   - JSONPath: .spec.forProvider.vpcId
     name: VPC
     type: string
+  - JSONPath: .spec.forProvider.cidrBlock
+    name: CIDR
+    type: string
   - JSONPath: .metadata.creationTimestamp
     name: AGE
     type: date
-  group: network.aws.crossplane.io
+  group: ec2.aws.crossplane.io
   names:
     categories:
     - crossplane
     - managed
     - aws
-    kind: InternetGateway
-    listKind: InternetGatewayList
-    plural: internetgateways
-    singular: internetgateway
+    kind: Subnet
+    listKind: SubnetList
+    plural: subnets
+    singular: subnet
   scope: Cluster
   subresources:
     status: {}
   validation:
     openAPIV3Schema:
-      description: An InternetGateway is a managed resource that represents an AWS
-        VPC Internet Gateway.
+      description: A Subnet is a managed resource that represents an AWS VPC Subnet.
       properties:
         apiVersion:
           description: 'APIVersion defines the versioned schema of this representation
@@ -55,7 +57,7 @@ spec:
         metadata:
           type: object
         spec:
-          description: An InternetGatewaySpec defines the desired state of an InternetGateway.
+          description: A SubnetSpec defines the desired state of a Subnet.
           properties:
             claimRef:
               description: ClaimReference specifies the resource claim to which this
@@ -135,9 +137,34 @@ spec:
                   type: string
               type: object
             forProvider:
-              description: InternetGatewayParameters define the desired state of an
-                AWS VPC Internet Gateway.
+              description: SubnetParameters define the desired state of an AWS VPC
+                Subnet.
               properties:
+                assignIpv6AddressOnCreation:
+                  description: Indicates whether a network interface created in this
+                    subnet (including a network interface created by RunInstances)
+                    receives an IPv6 address.
+                  type: boolean
+                availabilityZone:
+                  description: 'The Availability Zone for the subnet. Default: AWS
+                    selects one for you. If you create more than one subnet in your
+                    VPC, we may not necessarily select a different zone for each subnet.'
+                  type: string
+                availabilityZoneId:
+                  description: The AZ ID or the Local Zone ID of the subnet.
+                  type: string
+                cidrBlock:
+                  description: CIDRBlock is the IPv4 network range for the Subnet,
+                    in CIDR notation. For example, 10.0.0.0/18.
+                  type: string
+                ipv6CIDRBlock:
+                  description: The IPv6 network range for the subnet, in CIDR notation.
+                    The subnet size must use a /64 prefix length.
+                  type: string
+                mapPublicIPOnLaunch:
+                  description: Indicates whether instances launched in this subnet
+                    receive a public IPv4 address.
+                  type: boolean
                 tags:
                   description: Tags represents to current ec2 tags.
                   items:
@@ -158,7 +185,7 @@ spec:
                   description: VPCID is the ID of the VPC.
                   type: string
                 vpcIdRef:
-                  description: VPCIDRef references a VPC to and retrieves its vpcId
+                  description: VPCIDRef reference a VPC to retrieve its vpcId
                   properties:
                     name:
                       description: Name of the referenced object.
@@ -167,7 +194,7 @@ spec:
                   - name
                   type: object
                 vpcIdSelector:
-                  description: VPCIDSelector selects a reference to a VPC to and retrieves
+                  description: VPCIDSelector selects reference to a VPC to retrieve
                     its vpcId
                   properties:
                     matchControllerRef:
@@ -181,6 +208,8 @@ spec:
                         is selected.
                       type: object
                   type: object
+              required:
+              - cidrBlock
               type: object
             providerRef:
               description: ProviderReference specifies the provider that will be used
@@ -256,47 +285,29 @@ spec:
           - providerRef
           type: object
         status:
-          description: An InternetGatewayStatus represents the observed state of an
-            InternetGateway.
+          description: A SubnetStatus represents the observed state of a Subnet.
           properties:
             atProvider:
-              description: InternetGatewayObservation keeps the state for the external
-                resource
+              description: SubnetObservation keeps the state for the external resource
               properties:
-                attachments:
-                  description: Any VPCs attached to the internet gateway.
-                  items:
-                    description: InternetGatewayAttachment describes the attachment
-                      of a VPC to an internet gateway or an egress-only internet gateway.
-                    properties:
-                      attachmentStatus:
-                        description: The current state of the attachment. For an internet
-                          gateway, the state is available when attached to a VPC;
-                          otherwise, this value is not returned.
-                        enum:
-                        - available
-                        - attaching
-                        - attached
-                        - detaching
-                        - detached
-                        type: string
-                      vpcId:
-                        description: VPCID is the ID of the attached VPC.
-                        type: string
-                    required:
-                    - attachmentStatus
-                    - vpcId
-                    type: object
-                  type: array
-                internetGatewayId:
-                  description: The ID of the internet gateway.
+                availableIpAddressCount:
+                  description: The number of unused private IPv4 addresses in the
+                    subnet.
+                  format: int64
+                  type: integer
+                defaultForAz:
+                  description: Indicates whether this is the default subnet for the
+                    Availability Zone.
+                  type: boolean
+                subnetId:
+                  description: SubnetID is the ID of the Subnet.
                   type: string
-                ownerID:
-                  description: The ID of the AWS account that owns the internet gateway.
+                subnetState:
+                  description: SubnetState is the current state of the Subnet.
+                  enum:
+                  - pending
+                  - available
                   type: string
-              required:
-              - internetGatewayId
-              - ownerID
               type: object
             bindingPhase:
               description: Phase represents the binding phase of a managed resource
