+++ pkg/controller/aws.go
@@ -27,16 +27,16 @@ import (
 	"github.com/crossplane/provider-aws/pkg/controller/database"
 	"github.com/crossplane/provider-aws/pkg/controller/database/dbsubnetgroup"
 	"github.com/crossplane/provider-aws/pkg/controller/database/dynamodb"
+	"github.com/crossplane/provider-aws/pkg/controller/ec2/internetgateway"
+	"github.com/crossplane/provider-aws/pkg/controller/ec2/routetable"
+	"github.com/crossplane/provider-aws/pkg/controller/ec2/securitygroup"
+	"github.com/crossplane/provider-aws/pkg/controller/ec2/subnet"
+	"github.com/crossplane/provider-aws/pkg/controller/ec2/vpc"
 	"github.com/crossplane/provider-aws/pkg/controller/identity/iampolicy"
 	"github.com/crossplane/provider-aws/pkg/controller/identity/iamrole"
 	"github.com/crossplane/provider-aws/pkg/controller/identity/iamrolepolicyattachment"
 	"github.com/crossplane/provider-aws/pkg/controller/identity/iamuser"
 	"github.com/crossplane/provider-aws/pkg/controller/identity/iamuserpolicyattachment"
-	"github.com/crossplane/provider-aws/pkg/controller/network/internetgateway"
-	"github.com/crossplane/provider-aws/pkg/controller/network/routetable"
-	"github.com/crossplane/provider-aws/pkg/controller/network/securitygroup"
-	"github.com/crossplane/provider-aws/pkg/controller/network/subnet"
-	"github.com/crossplane/provider-aws/pkg/controller/network/vpc"
 	"github.com/crossplane/provider-aws/pkg/controller/s3"
 )
