+++ pkg/controller/ec2/vpc/controller_test.go
@@ -14,16 +14,20 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package subnet
+package vpc
 
 import (
 	"context"
 	"net/http"
 	"testing"
 
+	"github.com/crossplane/crossplane-runtime/pkg/resource"
+	"github.com/crossplane/crossplane-runtime/pkg/test"
+	"github.com/google/go-cmp/cmp"
+	"github.com/google/go-cmp/cmp/cmpopts"
+
 	"github.com/aws/aws-sdk-go-v2/aws"
 	awsec2 "github.com/aws/aws-sdk-go-v2/service/ec2"
-	"github.com/google/go-cmp/cmp"
 	"github.com/pkg/errors"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -33,9 +37,8 @@ import (
 	runtimev1alpha1 "github.com/crossplane/crossplane-runtime/apis/core/v1alpha1"
 	"github.com/crossplane/crossplane-runtime/pkg/meta"
 	"github.com/crossplane/crossplane-runtime/pkg/reconciler/managed"
-	"github.com/crossplane/crossplane-runtime/pkg/test"
 
-	"github.com/crossplane/provider-aws/apis/network/v1beta1"
+	"github.com/crossplane/provider-aws/apis/ec2/v1beta1"
 	awsv1alpha3 "github.com/crossplane/provider-aws/apis/v1alpha3"
 	awsclients "github.com/crossplane/provider-aws/pkg/clients"
 	"github.com/crossplane/provider-aws/pkg/clients/ec2"
@@ -53,38 +56,50 @@ const (
 )
 
 var (
-	subnetID = "some Id"
+	vpcID          = "some Id"
+	cidr           = "192.168.0.0/32"
+	tenancyDefault = "default"
 
 	errBoom = errors.New("boom")
 )
 
 type args struct {
-	subnet ec2.SubnetClient
-	kube   client.Client
-	cr     *v1beta1.Subnet
+	vpc  ec2.VPCClient
+	kube client.Client
+	cr   *v1beta1.VPC
 }
 
-type subnetModifier func(*v1beta1.Subnet)
+type vpcModifier func(*v1beta1.VPC)
 
-func withExternalName(name string) subnetModifier {
-	return func(r *v1beta1.Subnet) { meta.SetExternalName(r, name) }
+func withTags(tagMaps ...map[string]string) vpcModifier {
+	var tagList []v1beta1.Tag
+	for _, tagMap := range tagMaps {
+		for k, v := range tagMap {
+			tagList = append(tagList, v1beta1.Tag{Key: k, Value: v})
+		}
+	}
+	return func(r *v1beta1.VPC) { r.Spec.ForProvider.Tags = tagList }
 }
 
-func withConditions(c ...runtimev1alpha1.Condition) subnetModifier {
-	return func(r *v1beta1.Subnet) { r.Status.ConditionedStatus.Conditions = c }
+func withExternalName(name string) vpcModifier {
+	return func(r *v1beta1.VPC) { meta.SetExternalName(r, name) }
 }
 
-func withSpec(p v1beta1.SubnetParameters) subnetModifier {
-	return func(r *v1beta1.Subnet) { r.Spec.ForProvider = p }
+func withConditions(c ...runtimev1alpha1.Condition) vpcModifier {
+	return func(r *v1beta1.VPC) { r.Status.ConditionedStatus.Conditions = c }
 }
 
-func withStatus(s v1beta1.SubnetObservation) subnetModifier {
-	return func(r *v1beta1.Subnet) { r.Status.AtProvider = s }
+func withSpec(p v1beta1.VPCParameters) vpcModifier {
+	return func(r *v1beta1.VPC) { r.Spec.ForProvider = p }
 }
 
-func subnet(m ...subnetModifier) *v1beta1.Subnet {
-	cr := &v1beta1.Subnet{
-		Spec: v1beta1.SubnetSpec{
+func withStatus(s v1beta1.VPCObservation) vpcModifier {
+	return func(r *v1beta1.VPC) { r.Status.AtProvider = s }
+}
+
+func vpc(m ...vpcModifier) *v1beta1.VPC {
+	cr := &v1beta1.VPC{
+		Spec: v1beta1.VPCSpec{
 			ResourceSpec: runtimev1alpha1.ResourceSpec{
 				ProviderReference: &corev1.ObjectReference{Name: providerName},
 			},
@@ -129,8 +144,8 @@ func TestConnect(t *testing.T) {
 	}
 	type args struct {
 		kube        client.Client
-		newClientFn func(ctx context.Context, credentials []byte, region string, auth awsclients.AuthMethod) (ec2.SubnetClient, error)
-		cr          *v1beta1.Subnet
+		newClientFn func(ctx context.Context, credentials []byte, region string, auth awsclients.AuthMethod) (ec2.VPCClient, error)
+		cr          *v1beta1.VPC
 	}
 	type want struct {
 		err error
@@ -156,7 +171,7 @@ func TestConnect(t *testing.T) {
 						return errBoom
 					},
 				},
-				newClientFn: func(_ context.Context, credentials []byte, region string, _ awsclients.AuthMethod) (i ec2.SubnetClient, e error) {
+				newClientFn: func(_ context.Context, credentials []byte, region string, _ awsclients.AuthMethod) (i ec2.VPCClient, e error) {
 					if diff := cmp.Diff(credData, string(credentials)); diff != "" {
 						t.Errorf("r: -want, +got:\n%s", diff)
 					}
@@ -165,7 +180,7 @@ func TestConnect(t *testing.T) {
 					}
 					return nil, nil
 				},
-				cr: subnet(),
+				cr: vpc(),
 			},
 		},
 		"SuccessfulUseServiceAccount": {
@@ -180,7 +195,7 @@ func TestConnect(t *testing.T) {
 						return errBoom
 					},
 				},
-				newClientFn: func(_ context.Context, credentials []byte, region string, _ awsclients.AuthMethod) (i ec2.SubnetClient, e error) {
+				newClientFn: func(_ context.Context, credentials []byte, region string, _ awsclients.AuthMethod) (i ec2.VPCClient, e error) {
 					if diff := cmp.Diff("", string(credentials)); diff != "" {
 						t.Errorf("r: -want, +got:\n%s", diff)
 					}
@@ -189,7 +204,7 @@ func TestConnect(t *testing.T) {
 					}
 					return nil, nil
 				},
-				cr: subnet(),
+				cr: vpc(),
 			},
 		},
 		"ProviderGetFailed": {
@@ -199,7 +214,7 @@ func TestConnect(t *testing.T) {
 						return errBoom
 					},
 				},
-				cr: subnet(),
+				cr: vpc(),
 			},
 			want: want{
 				err: errors.Wrap(errBoom, errGetProvider),
@@ -221,7 +236,7 @@ func TestConnect(t *testing.T) {
 						}
 					},
 				},
-				cr: subnet(),
+				cr: vpc(),
 			},
 			want: want{
 				err: errors.Wrap(errBoom, errGetProviderSecret),
@@ -244,7 +259,7 @@ func TestConnect(t *testing.T) {
 						}
 					},
 				},
-				cr: subnet(),
+				cr: vpc(),
 			},
 			want: want{
 				err: errors.New(errGetProviderSecret),
@@ -254,7 +269,7 @@ func TestConnect(t *testing.T) {
 
 	for name, tc := range cases {
 		t.Run(name, func(t *testing.T) {
-			c := &connector{client: tc.kube, newClientFn: tc.newClientFn}
+			c := &connector{kube: tc.kube, newClientFn: tc.newClientFn}
 			_, err := c.Connect(context.Background(), tc.args.cr)
 			if diff := cmp.Diff(tc.err, err, test.EquateErrors()); diff != "" {
 				t.Errorf("r: -want, +got:\n%s", diff)
@@ -265,7 +280,7 @@ func TestConnect(t *testing.T) {
 
 func TestObserve(t *testing.T) {
 	type want struct {
-		cr     *v1beta1.Subnet
+		cr     *v1beta1.VPC
 		result managed.ExternalObservation
 		err    error
 	}
@@ -276,25 +291,41 @@ func TestObserve(t *testing.T) {
 	}{
 		"SuccessfulAvailable": {
 			args: args{
-				subnet: &fake.MockSubnetClient{
-					MockDescribe: func(input *awsec2.DescribeSubnetsInput) awsec2.DescribeSubnetsRequest {
-						return awsec2.DescribeSubnetsRequest{
-							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.DescribeSubnetsOutput{
-								Subnets: []awsec2.Subnet{
-									{
-										State: awsec2.SubnetStateAvailable,
-									},
-								},
+				kube: &test.MockClient{
+					MockUpdate: test.NewMockClient().Update,
+				},
+				vpc: &fake.MockVPCClient{
+					MockDescribe: func(input *awsec2.DescribeVpcsInput) awsec2.DescribeVpcsRequest {
+						return awsec2.DescribeVpcsRequest{
+							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.DescribeVpcsOutput{
+								Vpcs: []awsec2.Vpc{{
+									InstanceTenancy: awsec2.TenancyDefault,
+									State:           awsec2.VpcStateAvailable,
+								}},
+							}},
+						}
+					},
+					MockDescribeVpcAttributeRequest: func(input *awsec2.DescribeVpcAttributeInput) awsec2.DescribeVpcAttributeRequest {
+						return awsec2.DescribeVpcAttributeRequest{
+							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.DescribeVpcAttributeOutput{
+								EnableDnsHostnames: &awsec2.AttributeBooleanValue{},
+								EnableDnsSupport:   &awsec2.AttributeBooleanValue{},
 							}},
 						}
 					},
 				},
-				cr: subnet(withExternalName(subnetID)),
+				cr: vpc(withSpec(v1beta1.VPCParameters{
+					InstanceTenancy: aws.String(tenancyDefault),
+					CIDRBlock:       cidr,
+				}), withExternalName(vpcID)),
 			},
 			want: want{
-				cr: subnet(withStatus(v1beta1.SubnetObservation{
-					SubnetState: "available",
-				}), withExternalName(subnetID),
+				cr: vpc(withSpec(v1beta1.VPCParameters{
+					InstanceTenancy: aws.String(tenancyDefault),
+					CIDRBlock:       cidr,
+				}), withStatus(v1beta1.VPCObservation{
+					VPCState: "available",
+				}), withExternalName(vpcID),
 					withConditions(runtimev1alpha1.Available())),
 				result: managed.ExternalObservation{
 					ResourceExists:   true,
@@ -302,71 +333,55 @@ func TestObserve(t *testing.T) {
 				},
 			},
 		},
-		"MultipleSubnets": {
+		"MultipleVpcs": {
 			args: args{
-				subnet: &fake.MockSubnetClient{
-					MockDescribe: func(input *awsec2.DescribeSubnetsInput) awsec2.DescribeSubnetsRequest {
-						return awsec2.DescribeSubnetsRequest{
-							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.DescribeSubnetsOutput{
-								Subnets: []awsec2.Subnet{{}, {}},
+				kube: &test.MockClient{
+					MockUpdate: test.NewMockClient().Update,
+				},
+				vpc: &fake.MockVPCClient{
+					MockDescribe: func(input *awsec2.DescribeVpcsInput) awsec2.DescribeVpcsRequest {
+						return awsec2.DescribeVpcsRequest{
+							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.DescribeVpcsOutput{
+								Vpcs: []awsec2.Vpc{{}, {}},
 							}},
 						}
 					},
 				},
-				cr: subnet(withExternalName(subnetID)),
+				cr: vpc(withSpec(v1beta1.VPCParameters{
+					InstanceTenancy: aws.String(tenancyDefault),
+					CIDRBlock:       cidr,
+				}), withExternalName(vpcID)),
 			},
 			want: want{
-				cr:  subnet(withExternalName(subnetID)),
+				cr: vpc(withSpec(v1beta1.VPCParameters{
+					InstanceTenancy: aws.String(tenancyDefault),
+					CIDRBlock:       cidr,
+				}), withExternalName(vpcID)),
 				err: errors.New(errMultipleItems),
 			},
 		},
-		"NotUpToDate": {
+		"DescribeFail": {
 			args: args{
-				subnet: &fake.MockSubnetClient{
-					MockDescribe: func(input *awsec2.DescribeSubnetsInput) awsec2.DescribeSubnetsRequest {
-						return awsec2.DescribeSubnetsRequest{
-							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.DescribeSubnetsOutput{
-								Subnets: []awsec2.Subnet{
-									{
-										State:               awsec2.SubnetStateAvailable,
-										MapPublicIpOnLaunch: aws.Bool(false),
-									},
-								},
-							}},
-						}
-					},
+				kube: &test.MockClient{
+					MockUpdate: test.NewMockClient().Update,
 				},
-				cr: subnet(withSpec(v1beta1.SubnetParameters{
-					MapPublicIPOnLaunch: aws.Bool(true),
-				}),
-					withExternalName(subnetID)),
-			},
-			want: want{
-				cr: subnet(withSpec(v1beta1.SubnetParameters{
-					MapPublicIPOnLaunch: aws.Bool(true),
-				}), withStatus(v1beta1.SubnetObservation{
-					SubnetState: string(awsec2.SubnetStateAvailable),
-				}), withExternalName(subnetID),
-					withConditions(runtimev1alpha1.Available())),
-				result: managed.ExternalObservation{
-					ResourceExists:   true,
-					ResourceUpToDate: false,
-				},
-			},
-		},
-		"DescribeFailed": {
-			args: args{
-				subnet: &fake.MockSubnetClient{
-					MockDescribe: func(input *awsec2.DescribeSubnetsInput) awsec2.DescribeSubnetsRequest {
-						return awsec2.DescribeSubnetsRequest{
+				vpc: &fake.MockVPCClient{
+					MockDescribe: func(input *awsec2.DescribeVpcsInput) awsec2.DescribeVpcsRequest {
+						return awsec2.DescribeVpcsRequest{
 							Request: &aws.Request{HTTPRequest: &http.Request{}, Error: errBoom},
 						}
 					},
 				},
-				cr: subnet(withExternalName(subnetID)),
+				cr: vpc(withSpec(v1beta1.VPCParameters{
+					InstanceTenancy: aws.String(tenancyDefault),
+					CIDRBlock:       cidr,
+				}), withExternalName(vpcID)),
 			},
 			want: want{
-				cr:  subnet(withExternalName(subnetID)),
+				cr: vpc(withSpec(v1beta1.VPCParameters{
+					InstanceTenancy: aws.String(tenancyDefault),
+					CIDRBlock:       cidr,
+				}), withExternalName(vpcID)),
 				err: errors.Wrap(errBoom, errDescribe),
 			},
 		},
@@ -374,7 +389,7 @@ func TestObserve(t *testing.T) {
 
 	for name, tc := range cases {
 		t.Run(name, func(t *testing.T) {
-			e := &external{kube: tc.kube, client: tc.subnet}
+			e := &external{kube: tc.kube, client: tc.vpc}
 			o, err := e.Observe(context.Background(), tc.args.cr)
 
 			if diff := cmp.Diff(tc.want.err, err, test.EquateErrors()); diff != "" {
@@ -392,7 +407,7 @@ func TestObserve(t *testing.T) {
 
 func TestCreate(t *testing.T) {
 	type want struct {
-		cr     *v1beta1.Subnet
+		cr     *v1beta1.VPC
 		result managed.ExternalCreation
 		err    error
 	}
@@ -407,41 +422,47 @@ func TestCreate(t *testing.T) {
 					MockUpdate:       test.NewMockClient().Update,
 					MockStatusUpdate: test.NewMockClient().MockStatusUpdate,
 				},
-				subnet: &fake.MockSubnetClient{
-					MockCreate: func(input *awsec2.CreateSubnetInput) awsec2.CreateSubnetRequest {
-						return awsec2.CreateSubnetRequest{
-							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.CreateSubnetOutput{
-								Subnet: &awsec2.Subnet{
-									SubnetId: aws.String(subnetID),
+				vpc: &fake.MockVPCClient{
+					MockCreate: func(input *awsec2.CreateVpcInput) awsec2.CreateVpcRequest {
+						return awsec2.CreateVpcRequest{
+							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.CreateVpcOutput{
+								Vpc: &awsec2.Vpc{
+									VpcId:     aws.String(vpcID),
+									CidrBlock: aws.String(cidr),
 								},
 							}},
 						}
 					},
+					MockModifyAttribute: func(input *awsec2.ModifyVpcAttributeInput) awsec2.ModifyVpcAttributeRequest {
+						return awsec2.ModifyVpcAttributeRequest{
+							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.ModifyVpcAttributeOutput{}},
+						}
+					},
 				},
-				cr: subnet(),
+				cr: vpc(),
 			},
 			want: want{
-				cr: subnet(withExternalName(subnetID),
+				cr: vpc(withExternalName(vpcID),
 					withConditions(runtimev1alpha1.Creating())),
 			},
 		},
-		"CreateFailed": {
+		"CreateFail": {
 			args: args{
 				kube: &test.MockClient{
 					MockUpdate:       test.NewMockClient().Update,
 					MockStatusUpdate: test.NewMockClient().MockStatusUpdate,
 				},
-				subnet: &fake.MockSubnetClient{
-					MockCreate: func(input *awsec2.CreateSubnetInput) awsec2.CreateSubnetRequest {
-						return awsec2.CreateSubnetRequest{
+				vpc: &fake.MockVPCClient{
+					MockCreate: func(input *awsec2.CreateVpcInput) awsec2.CreateVpcRequest {
+						return awsec2.CreateVpcRequest{
 							Request: &aws.Request{HTTPRequest: &http.Request{}, Error: errBoom},
 						}
 					},
 				},
-				cr: subnet(),
+				cr: vpc(),
 			},
 			want: want{
-				cr:  subnet(withConditions(runtimev1alpha1.Creating())),
+				cr:  vpc(withConditions(runtimev1alpha1.Creating())),
 				err: errors.Wrap(errBoom, errCreate),
 			},
 		},
@@ -449,7 +470,7 @@ func TestCreate(t *testing.T) {
 
 	for name, tc := range cases {
 		t.Run(name, func(t *testing.T) {
-			e := &external{kube: tc.kube, client: tc.subnet}
+			e := &external{kube: tc.kube, client: tc.vpc}
 			o, err := e.Create(context.Background(), tc.args.cr)
 
 			if diff := cmp.Diff(tc.want.err, err, test.EquateErrors()); diff != "" {
@@ -467,7 +488,7 @@ func TestCreate(t *testing.T) {
 
 func TestUpdate(t *testing.T) {
 	type want struct {
-		cr     *v1beta1.Subnet
+		cr     *v1beta1.VPC
 		result managed.ExternalUpdate
 		err    error
 	}
@@ -478,75 +499,68 @@ func TestUpdate(t *testing.T) {
 	}{
 		"Successful": {
 			args: args{
-				subnet: &fake.MockSubnetClient{
-					MockModify: func(input *awsec2.ModifySubnetAttributeInput) awsec2.ModifySubnetAttributeRequest {
-						return awsec2.ModifySubnetAttributeRequest{
-							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.ModifySubnetAttributeOutput{}},
+				vpc: &fake.MockVPCClient{
+					MockModifyTenancy: func(input *awsec2.ModifyVpcTenancyInput) awsec2.ModifyVpcTenancyRequest {
+						return awsec2.ModifyVpcTenancyRequest{
+							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.ModifyVpcTenancyOutput{}},
 						}
 					},
-					MockDescribe: func(input *awsec2.DescribeSubnetsInput) awsec2.DescribeSubnetsRequest {
-						return awsec2.DescribeSubnetsRequest{
-							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.DescribeSubnetsOutput{
-								Subnets: []awsec2.Subnet{{
-									SubnetId:            aws.String(subnetID),
-									MapPublicIpOnLaunch: aws.Bool(false),
-								}},
-							}},
+					MockCreateTagsRequest: func(input *awsec2.CreateTagsInput) awsec2.CreateTagsRequest {
+						return awsec2.CreateTagsRequest{
+							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.CreateTagsOutput{}},
+						}
+					},
+					MockModifyAttribute: func(input *awsec2.ModifyVpcAttributeInput) awsec2.ModifyVpcAttributeRequest {
+						return awsec2.ModifyVpcAttributeRequest{
+							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.ModifyVpcAttributeOutput{}},
 						}
 					},
 				},
-				cr: subnet(withSpec(v1beta1.SubnetParameters{
-					MapPublicIPOnLaunch: aws.Bool(true),
-				}), withStatus(v1beta1.SubnetObservation{
-					SubnetID: subnetID,
+				cr: vpc(withSpec(v1beta1.VPCParameters{
+					InstanceTenancy: aws.String(tenancyDefault),
 				})),
 			},
 			want: want{
-				cr: subnet(withSpec(v1beta1.SubnetParameters{
-					MapPublicIPOnLaunch: aws.Bool(true),
-				}), withStatus(v1beta1.SubnetObservation{
-					SubnetID: subnetID,
+				cr: vpc(withSpec(v1beta1.VPCParameters{
+					InstanceTenancy: aws.String(tenancyDefault),
 				})),
 			},
 		},
 		"ModifyFailed": {
 			args: args{
-				subnet: &fake.MockSubnetClient{
-					MockModify: func(input *awsec2.ModifySubnetAttributeInput) awsec2.ModifySubnetAttributeRequest {
-						return awsec2.ModifySubnetAttributeRequest{
-							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.ModifySubnetAttributeOutput{}},
+				vpc: &fake.MockVPCClient{
+					MockModifyTenancy: func(input *awsec2.ModifyVpcTenancyInput) awsec2.ModifyVpcTenancyRequest {
+						return awsec2.ModifyVpcTenancyRequest{
+							Request: &aws.Request{HTTPRequest: &http.Request{}, Error: errBoom},
 						}
 					},
-					MockDescribe: func(input *awsec2.DescribeSubnetsInput) awsec2.DescribeSubnetsRequest {
-						return awsec2.DescribeSubnetsRequest{
-							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.DescribeSubnetsOutput{
-								Subnets: []awsec2.Subnet{{
-									SubnetId:            aws.String(subnetID),
-									MapPublicIpOnLaunch: aws.Bool(false),
-								}},
-							}},
+					MockCreateTagsRequest: func(input *awsec2.CreateTagsInput) awsec2.CreateTagsRequest {
+						return awsec2.CreateTagsRequest{
+							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.CreateTagsOutput{}},
+						}
+					},
+					MockModifyAttribute: func(input *awsec2.ModifyVpcAttributeInput) awsec2.ModifyVpcAttributeRequest {
+						return awsec2.ModifyVpcAttributeRequest{
+							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.ModifyVpcAttributeOutput{}},
 						}
 					},
 				},
-				cr: subnet(withSpec(v1beta1.SubnetParameters{
-					MapPublicIPOnLaunch: aws.Bool(true),
-				}), withStatus(v1beta1.SubnetObservation{
-					SubnetID: subnetID,
+				cr: vpc(withSpec(v1beta1.VPCParameters{
+					InstanceTenancy: aws.String(tenancyDefault),
 				})),
 			},
 			want: want{
-				cr: subnet(withSpec(v1beta1.SubnetParameters{
-					MapPublicIPOnLaunch: aws.Bool(true),
-				}), withStatus(v1beta1.SubnetObservation{
-					SubnetID: subnetID,
+				cr: vpc(withSpec(v1beta1.VPCParameters{
+					InstanceTenancy: aws.String(tenancyDefault),
 				})),
+				err: errors.Wrap(errBoom, errUpdate),
 			},
 		},
 	}
 
 	for name, tc := range cases {
 		t.Run(name, func(t *testing.T) {
-			e := &external{kube: tc.kube, client: tc.subnet}
+			e := &external{kube: tc.kube, client: tc.vpc}
 			u, err := e.Update(context.Background(), tc.args.cr)
 
 			if diff := cmp.Diff(tc.want.err, err, test.EquateErrors()); diff != "" {
@@ -564,7 +578,7 @@ func TestUpdate(t *testing.T) {
 
 func TestDelete(t *testing.T) {
 	type want struct {
-		cr  *v1beta1.Subnet
+		cr  *v1beta1.VPC
 		err error
 	}
 
@@ -574,40 +588,32 @@ func TestDelete(t *testing.T) {
 	}{
 		"Successful": {
 			args: args{
-				subnet: &fake.MockSubnetClient{
-					MockDelete: func(input *awsec2.DeleteSubnetInput) awsec2.DeleteSubnetRequest {
-						return awsec2.DeleteSubnetRequest{
-							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.DeleteSubnetOutput{}},
+				vpc: &fake.MockVPCClient{
+					MockDelete: func(input *awsec2.DeleteVpcInput) awsec2.DeleteVpcRequest {
+						return awsec2.DeleteVpcRequest{
+							Request: &aws.Request{HTTPRequest: &http.Request{}, Data: &awsec2.DeleteVpcOutput{}},
 						}
 					},
 				},
-				cr: subnet(withStatus(v1beta1.SubnetObservation{
-					SubnetID: subnetID,
-				})),
+				cr: vpc(),
 			},
 			want: want{
-				cr: subnet(withStatus(v1beta1.SubnetObservation{
-					SubnetID: subnetID,
-				}), withConditions(runtimev1alpha1.Deleting())),
+				cr: vpc(withConditions(runtimev1alpha1.Deleting())),
 			},
 		},
 		"DeleteFailed": {
 			args: args{
-				subnet: &fake.MockSubnetClient{
-					MockDelete: func(input *awsec2.DeleteSubnetInput) awsec2.DeleteSubnetRequest {
-						return awsec2.DeleteSubnetRequest{
+				vpc: &fake.MockVPCClient{
+					MockDelete: func(input *awsec2.DeleteVpcInput) awsec2.DeleteVpcRequest {
+						return awsec2.DeleteVpcRequest{
 							Request: &aws.Request{HTTPRequest: &http.Request{}, Error: errBoom},
 						}
 					},
 				},
-				cr: subnet(withStatus(v1beta1.SubnetObservation{
-					SubnetID: subnetID,
-				})),
+				cr: vpc(),
 			},
 			want: want{
-				cr: subnet(withStatus(v1beta1.SubnetObservation{
-					SubnetID: subnetID,
-				}), withConditions(runtimev1alpha1.Deleting())),
+				cr:  vpc(withConditions(runtimev1alpha1.Deleting())),
 				err: errors.Wrap(errBoom, errDelete),
 			},
 		},
@@ -615,13 +621,62 @@ func TestDelete(t *testing.T) {
 
 	for name, tc := range cases {
 		t.Run(name, func(t *testing.T) {
-			e := &external{kube: tc.kube, client: tc.subnet}
+			e := &external{kube: tc.kube, client: tc.vpc}
 			err := e.Delete(context.Background(), tc.args.cr)
 
 			if diff := cmp.Diff(tc.want.err, err, test.EquateErrors()); diff != "" {
 				t.Errorf("r: -want, +got:\n%s", diff)
 			}
 			if diff := cmp.Diff(tc.want.cr, tc.args.cr, test.EquateConditions()); diff != "" {
+				t.Errorf("r: -want, +got:\n%s", diff)
+			}
+		})
+	}
+}
+
+func TestInitialize(t *testing.T) {
+	type args struct {
+		cr   *v1beta1.VPC
+		kube client.Client
+	}
+	type want struct {
+		cr  *v1beta1.VPC
+		err error
+	}
+
+	cases := map[string]struct {
+		args
+		want
+	}{
+		"Successful": {
+			args: args{
+				cr:   vpc(withTags(map[string]string{"foo": "bar"})),
+				kube: &test.MockClient{MockUpdate: test.NewMockUpdateFn(nil)},
+			},
+			want: want{
+				cr: vpc(withTags(resource.GetExternalTags(vpc()), map[string]string{"foo": "bar"})),
+			},
+		},
+		"UpdateFailed": {
+			args: args{
+				cr:   vpc(),
+				kube: &test.MockClient{MockUpdate: test.NewMockUpdateFn(errBoom)},
+			},
+			want: want{
+				err: errors.Wrap(errBoom, errKubeUpdateFailed),
+			},
+		},
+	}
+
+	for name, tc := range cases {
+		t.Run(name, func(t *testing.T) {
+			e := &tagger{kube: tc.kube}
+			err := e.Initialize(context.Background(), tc.args.cr)
+
+			if diff := cmp.Diff(tc.want.err, err, test.EquateErrors()); diff != "" {
+				t.Errorf("r: -want, +got:\n%s", diff)
+			}
+			if diff := cmp.Diff(tc.want.cr, tc.args.cr, cmpopts.SortSlices(func(a, b v1beta1.Tag) bool { return a.Key > b.Key })); err == nil && diff != "" {
 				t.Errorf("r: -want, +got:\n%s", diff)
 			}
 		})
